[feat]: add command description && move runHostGroup command
